Added error handling in case the OpenAPI calls fail

+++ api/controllers/set.js
@@ -75,27 +75,47 @@ router.post(
 			}
 		}
 
-		for (let Term_Id of Object.keys(data)) {
-			let summary = await summarizer.summarize(data[Term_Id][0], data[Term_Id][1]);
+		try {
+			for (let Term_Id of Object.keys(data)) {
+				let summary = await summarizer.summarize(
+					data[Term_Id][0],
+					data[Term_Id][1]
+				);
 
-			await Summary.destroy({
-				where: { Term_Id: Term_Id },
-			}).catch((error) => {
-				console.log(error);
-				res.status(400).json({
-					message: "Failed to remove existing summary",
-				});
-			});
+				if (summary instanceof OpenAI.APIError) {
+					console.log(`OpenAI: ${err.name} ${err.status}`);
+					throw err;
+				}
 
-			await Summary.create({
-				Term_Id: Term_Id,
-				Content: summary,
-			}).catch((error) => {
-				console.log(error);
-				res.status(400).json({
-					message: "Error making summary",
+				await Summary.destroy({
+					where: { Term_Id: Term_Id },
+				}).catch((error) => {
+					console.log(error);
+					res.status(400).json({
+						message: "Failed to remove existing summary",
+					});
 				});
-			});
+
+				await Summary.create({
+					Term_Id: Term_Id,
+					Content: summary,
+				}).catch((error) => {
+					console.log(error);
+					res.status(400).json({
+						message: "Error making summary",
+					});
+				});
+			}
+		} catch (error) {
+			if (error.status == 429)
+				res.status(500).json({
+					message: "OpenAI - ran out of credits",
+				});
+			else
+				res.status(500).json({
+					message: "Could not obtain response from the OpenAI API",
+				});
+			return;
 		}
 
 		// Get final output to send back
